requirements.txt
python-dotenv==1.0.0
PyQt5==5.15.10
